Add DTO in loginRequestDTO.php which is used in login.php and also reorganize composer.json.
