+++ public/index.php
@@ -1,7 +1,7 @@
 <?php
 session_start();
 
-use App\messageHandler;
+use App\classes\messageHandler;
 
 require_once __DIR__ . '/../vendor/autoload.php';
